Bootstrap, JS, and carousel on Play page
Added Botstrap and used it to add an image carousel to the bottom of the play page. Since there is now external CSS and JS, page load time may slow down noticeably. I'm mainly using this as a placeholder until I can learn how to do this on my own.